feat(docs): wip
